+++ repository/src/main/java/com/sj/repository/service/ProviderService.java
@@ -3,5 +3,5 @@ package com.sj.repository.service;
 import com.sj.model.model.Provider;
 
 public interface ProviderService {
-	public Provider findOne(int id);
+	public Provider findOne(Long id);
 }
